requirements.txt
# Core Flask application dependencies
flask==2.3.3
yt-dlp==2023.9.24
opencv-python==4.8.1.78
pillow==10.0.1
requests==2.31.0
werkzeug==2.3.7
moviepy==1.0.3
google-api-python-client==2.100.0
python-dotenv==1.0.0

# YouTube Upload dependencies
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
google-api-core==2.24.2
uritemplate==4.2.0
cachetools==5.5.2
rsa==4.9
pyasn1-modules==0.4.1
requests-oauthlib==2.0.0
oauthlib==3.2.2

# Enhanced functionality dependencies
flask-limiter==3.12
gunicorn==23.0.0
limits>=3.13
ordered-set>=4.1.0
rich>=12.0.0

# Testing dependencies
pytest==8.4.1
pytest-flask==1.3.0
pytest-mock==3.14.1
packaging>=25.0

# Performance monitoring
psutil==7.0.0

# Production database/cache support
redis==6.4.0
async-timeout>=4.0.3
typing-extensions>=4.6.0

# Additional utility dependencies
deprecated>=1.2.0
markdown-it-py>=2.2.0
mdurl>=0.1.0
wrapt>=1.10.0
pygments>=2.7.2
exceptiongroup>=1.0.0
tomli>=1.0.0
iniconfig>=1.0.0
pluggy>=1.5.0
